refactor contactPane.js, split out toolPane.js, debugging group-only working, not tested
